# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/bank_statement_imports:
    get:
      tags:
        - BankStatementImports
      summary: List BankStatementImports
      description: Get a paginated list of BankStatementImports with optional filtering and sorting
      operationId: listBankStatementImports
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankStatementImportList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - BankStatementImports
      summary: Create BankStatementImport
      description: Create a new BankStatementImport
      operationId: createBankStatementImport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBankStatementImportRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankStatementImport'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/bank_transactions:
    get:
      tags:
        - BankTransactions
      summary: List BankTransactions
      description: Get a paginated list of BankTransactions with optional filtering and sorting
      operationId: listBankTransactions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankTransactionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - BankTransactions
      summary: Create BankTransaction
      description: Create a new BankTransaction
      operationId: createBankTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBankTransactionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankTransaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    BankStatementImport:
      type: object
      required:
        - id
        - company_id
        - bank_account_id
        - source_format
        - statement_period_start
        - statement_period_end
        - opening_balance
        - closing_balance
        - status
        - row_count
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        bank_account_id:
          type: string
          format: uuid
        source_format:
          type: string
        statement_period_start:
          type: string
          format: date
        statement_period_end:
          type: string
          format: date
        opening_balance:
          type: string
        closing_balance:
          type: string
        file_ref:
          nullable: true
          type: string
          maxLength: 300
        status:
          type: string
        row_count:
          type: integer
          format: int32
        metadata:
          type: object
...
    CreateBankStatementImportRequest:
      type: object
      required:
        - company_id
        - bank_account_id
        - source_format
        - statement_period_start
        - statement_period_end
        - opening_balance
        - closing_balance
        - status
        - row_count
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        bank_account_id:
          type: string
          format: uuid
        source_format:
          type: string
        statement_period_start:
          type: string
          format: date
        statement_period_end:
          type: string
          format: date
        opening_balance:
          type: string
        closing_balance:
          type: string
        file_ref:
          nullable: true
          type: string
          maxLength: 300
        status:
          type: string
        row_count:
          type: integer
          format: int32
        metadata:
          type: object
...
    BankStatementImportList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BankStatementImport'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    BankTransaction:
      type: object
      required:
        - id
        - company_id
        - bank_account_id
        - import_id
        - txn_date
        - deposit
        - withdrawal
        - currency
        - status
        - allocated_amount
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        bank_account_id:
          type: string
          format: uuid
        import_id:
          type: string
          format: uuid
        txn_date:
          type: string
          format: date
        value_date:
          nullable: true
          type: string
          format: date
        description:
          nullable: true
          type: string
          maxLength: 300
        reference_no:
          nullable: true
          type: string
          maxLength: 140
        deposit:
          type: string
        withdrawal:
          type: string
        currency:
          type: string
          maxLength: 3
        status:
          type: string
        allocated_amount:
          type: string
        metadata:
          type: object
...
    CreateBankTransactionRequest:
      type: object
      required:
        - company_id
        - bank_account_id
        - import_id
        - txn_date
        - deposit
        - withdrawal
        - currency
        - status
        - allocated_amount
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        bank_account_id:
          type: string
          format: uuid
        import_id:
          type: string
          format: uuid
        txn_date:
          type: string
          format: date
        value_date:
          nullable: true
          type: string
          format: date
        description:
          nullable: true
          type: string
          maxLength: 300
        reference_no:
          nullable: true
          type: string
          maxLength: 140
        deposit:
          type: string
        withdrawal:
          type: string
        currency:
          type: string
          maxLength: 3
        status:
          type: string
        allocated_amount:
          type: string
        metadata:
          type: object
...
    BankTransactionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BankTransaction'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []