# src/presentation/http/bank_statement_import_handler.rs
//! BankStatementImport REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{NaiveDate};
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{BankStatementImportService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateBankStatementImportDto, UpdateBankStatementImportDto, PatchBankStatementImportDto, BankStatementImportResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum BankStatementImportError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for BankStatementImportError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for BankStatementImportError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "BANKSTATEMENTIMPORT_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "BANKSTATEMENTIMPORT_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "BANKSTATEMENTIMPORT_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "BANKSTATEMENTIMPORT_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for BankStatementImport.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /bank_statement_imports | List with pagination |
/// | POST | /bank_statement_imports | Create new |
/// | GET | /bank_statement_imports/:id | Get by ID |
/// | PUT | /bank_statement_imports/:id | Full update |
/// | PATCH | /bank_statement_imports/:id | Partial update |
/// | DELETE | /bank_statement_imports/:id | Soft delete |
/// | POST | /bank_statement_imports/bulk | Bulk create |
/// | POST | /bank_statement_imports/upsert | Upsert |
/// | GET | /bank_statement_imports/trash | List deleted |
/// | POST | /bank_statement_imports/:id/restore | Restore |
/// | DELETE | /bank_statement_imports/empty | Empty trash |
/// | GET | /bank_statement_imports/:id/deleted | Get deleted by ID |
/// | DELETE | /bank_statement_imports/trash/:id | Permanent delete from trash |
/// | GET | /bank_statement_imports/count | Count active entities |
/// | GET | /bank_statement_imports/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(BankStatementImportService::with_repository(repository));
/// let router = create_bank_statement_import_routes(service);
/// ```
pub fn create_bank_statement_import_routes(service: Arc<BankStatementImportService>) -> Router {
    BackboneCrudHandler::<BankStatementImportService, BankStatementImport, CreateBankStatementImportDto, UpdateBankStatementImportDto, BankStatementImportResponseDto>::routes(
        service,
        "/bank_statement_imports",
    )
}

/// Create Axum router with only the read (GET) endpoints for BankStatementImport.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_bank_statement_import_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_bank_statement_import_read_routes(service: Arc<BankStatementImportService>) -> Router {
    BackboneCrudHandler::<BankStatementImportService, BankStatementImport, CreateBankStatementImportDto, UpdateBankStatementImportDto, BankStatementImportResponseDto>::read_routes(
        service,
        "/bank_statement_imports",
    )
}

/// Create Axum router with only the write (mutation) endpoints for BankStatementImport.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_bank_statement_import_write_routes(service: Arc<BankStatementImportService>) -> Router {
    BackboneCrudHandler::<BankStatementImportService, BankStatementImport, CreateBankStatementImportDto, UpdateBankStatementImportDto, BankStatementImportResponseDto>::write_routes(
        service,
        "/bank_statement_imports",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_bank_statement_import_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<BankStatementImportService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_bank_statement_import_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}