# src/presentation/http/bank_transaction_handler.rs
//! BankTransaction REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{NaiveDate};
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{BankTransactionService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateBankTransactionDto, UpdateBankTransactionDto, PatchBankTransactionDto, BankTransactionResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum BankTransactionError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for BankTransactionError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for BankTransactionError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "BANKTRANSACTION_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "BANKTRANSACTION_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "BANKTRANSACTION_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "BANKTRANSACTION_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for BankTransaction.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /bank_transactions | List with pagination |
/// | POST | /bank_transactions | Create new |
/// | GET | /bank_transactions/:id | Get by ID |
/// | PUT | /bank_transactions/:id | Full update |
/// | PATCH | /bank_transactions/:id | Partial update |
/// | DELETE | /bank_transactions/:id | Soft delete |
/// | POST | /bank_transactions/bulk | Bulk create |
/// | POST | /bank_transactions/upsert | Upsert |
/// | GET | /bank_transactions/trash | List deleted |
/// | POST | /bank_transactions/:id/restore | Restore |
/// | DELETE | /bank_transactions/empty | Empty trash |
/// | GET | /bank_transactions/:id/deleted | Get deleted by ID |
/// | DELETE | /bank_transactions/trash/:id | Permanent delete from trash |
/// | GET | /bank_transactions/count | Count active entities |
/// | GET | /bank_transactions/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(BankTransactionService::with_repository(repository));
/// let router = create_bank_transaction_routes(service);
/// ```
pub fn create_bank_transaction_routes(service: Arc<BankTransactionService>) -> Router {
    BackboneCrudHandler::<BankTransactionService, BankTransaction, CreateBankTransactionDto, UpdateBankTransactionDto, BankTransactionResponseDto>::routes(
        service,
        "/bank_transactions",
    )
}

/// Create Axum router with only the read (GET) endpoints for BankTransaction.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_bank_transaction_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_bank_transaction_read_routes(service: Arc<BankTransactionService>) -> Router {
    BackboneCrudHandler::<BankTransactionService, BankTransaction, CreateBankTransactionDto, UpdateBankTransactionDto, BankTransactionResponseDto>::read_routes(
        service,
        "/bank_transactions",
    )
}

/// Create Axum router with only the write (mutation) endpoints for BankTransaction.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_bank_transaction_write_routes(service: Arc<BankTransactionService>) -> Router {
    BackboneCrudHandler::<BankTransactionService, BankTransaction, CreateBankTransactionDto, UpdateBankTransactionDto, BankTransactionResponseDto>::write_routes(
        service,
        "/bank_transactions",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_bank_transaction_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<BankTransactionService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_bank_transaction_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}